--- src/main/java/com/marginallyclever/ro3/apps/viewport/renderpass/RenderPassHelper.java
@@ -1,4 +1,0 @@
-package com.marginallyclever.ro3.apps.viewport.renderpass;
-
-public class RenderPassHelper {
-}
